requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6
sse-starlette==1.8.2
pyyaml==6.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
aiofiles==23.2.1
sqlalchemy==2.0.25
aiosqlite==0.19.0
langgraph==1.0.20
mido==1.3.3
fabric==3.2.2
python-rtmidi>=1.5.0
httpx==0.27.0
